DEV: Fix compatibility with the Glimmer Post Stream (#31)
This change is needed because the TC handles the DOM elements once the component is inserted and while the PostStream widget isn't modernized, the PostStream components are also performing the same task to include classes in some components.

Unfortunately, this was causing the TC to try to get properties from a selector that wasn't ready yet in the DOM.

Once we have a Glimmer component version of the PostStream widget, we will be able to revert the change.